+++ packages/editor/src/RichContentEditor/Toolbars/createEditorToolbars.js
@@ -1,7 +1,6 @@
-import { createSideToolbar } from './SideToolbar';
-import { createMobileToolbar, createFooterToolbar, createStaticTextToolbar } from './StaticToolbar';
-import { createInlineTextToolbar } from './InlineToolbar';
-import { simplePubsub, getToolbarTheme, WixUtils } from 'wix-rich-content-common';
+import merge from 'lodash/merge';
+import { simplePubsub, getToolbarTheme, getConfigByFormFactor } from 'wix-rich-content-common';
+import { getDefaultToolbarSettings } from './default-toolbar-settings';
 
 const createEditorToolbars = config => {
   const {
@@ -9,96 +8,47 @@ const createEditorToolbars = config => {
     anchorTarget,
     relValue,
     // textToolbarType,
-    alwaysShowSideToolbar,
-    hideFooterToolbar,
-    sideToolbarOffset,
+    textAlignment,
     helpers,
     isMobile,
     theme,
     getEditorState,
     setEditorState,
-    textAlignment,
     t,
-    refId
+    refId,
+    getToolbarSettings = () => {}
   } = config;
   const { pluginButtons, pluginTextButtonMappers, textButtons } = buttons;
-  const sideToolbarPluginButtons = pluginButtons && pluginButtons
-    .filter(({ originalConfig }) => originalConfig.addToSideToolbar !== false)
-    .map(({ component }) => component);
-  const footerToolbarPluginButtons = pluginButtons && pluginButtons
-    .filter(({ originalConfig }) => originalConfig.addToFooterToolbar !== false)
-    .map(({ component }) => component);
+
   const pubsub = simplePubsub();
-  const shouldCreateSidePluginToolbar = pluginButtons && sideToolbarPluginButtons.length;
-  const shouldCreateFooterPluginToolbar = pluginButtons && footerToolbarPluginButtons.length;
-  const shouldCreateTextToolbar = !isMobile || WixUtils.isiOS();
+
+  const defaultToolbarSettings = getDefaultToolbarSettings({ isMobile, pluginButtons, textButtons });
+  const customSettings = getToolbarSettings({ isMobile, pluginButtons, textButtons });
+
+  const toolbarSettings = merge(defaultToolbarSettings, customSettings);
 
   const toolbars = {};
-  if (shouldCreateSidePluginToolbar) {
-    toolbars.side = createSideToolbar({
-      refId,
-      buttons: sideToolbarPluginButtons,
-      alwaysShow: alwaysShowSideToolbar,
-      offset: sideToolbarOffset,
-      theme: { ...getToolbarTheme(theme, 'side'), ...theme },
-      pubsub,
-      isMobile,
-      helpers,
-      t
-    });
-  }
 
-  if (shouldCreateTextToolbar) {
-    toolbars.textInline = createInlineTextToolbar({
-      refId,
-      buttons: textButtons,
-      pluginTextButtonMappers,
-      defaultTextAlignment: textAlignment,
-      theme: { ...getToolbarTheme(theme, 'inline'), ...theme },
-      anchorTarget,
-      relValue,
-      pubsub,
-      isMobile,
-      helpers,
-      t,
-    });
-    toolbars.textStatic = createStaticTextToolbar({
-      refId,
-      buttons: textButtons,
-      pluginTextButtonMappers,
-      theme: { ...getToolbarTheme(theme, 'text'), ...theme },
-      anchorTarget,
-      relValue,
-      pubsub,
-      isMobile,
-      helpers,
-      t,
-    });
-  }
-
-  if (!isMobile) {
-    if (shouldCreateFooterPluginToolbar && !hideFooterToolbar) {
-      toolbars.footer = createFooterToolbar({
+  toolbarSettings.filter(({ shouldCreate }) => getConfigByFormFactor({ config: shouldCreate(), isMobile, defaultValue: true }))
+    .forEach(({ name, getButtons, getPositionOffset, getVisibilityFn, getInstance }) => {
+      toolbars[name] = getInstance({
+        buttons: getConfigByFormFactor({ config: getButtons(), isMobile, defaultValue: [] }),
+        offset: getConfigByFormFactor({ config: getPositionOffset(), isMobile, defaultValue: { x: 0, y: 0 } }),
+        visibilityFn: getConfigByFormFactor({ config: getVisibilityFn(), isMobile, defaultValue: () => true }),
+        theme: { ...getToolbarTheme(theme, name.toLowerCase()), ...theme },
+        defaultTextAlignment: textAlignment,
+        pluginTextButtonMappers,
+        getEditorState,
+        setEditorState,
+        anchorTarget,
+        relValue,
+        isMobile,
+        helpers,
+        pubsub,
         refId,
-        buttons: footerToolbarPluginButtons,
-        theme: { ...getToolbarTheme(theme, 'footer'), ...theme },
+        t
       });
-    }
-  } else {
-    toolbars.mobile = createMobileToolbar({
-      refId,
-      anchorTarget,
-      relValue,
-      buttons,
-      pluginTextButtonMappers,
-      helpers,
-      pubsub,
-      getEditorState,
-      setEditorState,
-      t,
-      theme: { ...getToolbarTheme(theme, 'mobile'), ...theme },
     });
-  }
 
   return toolbars;
 };