+++ packages/plugin-image/src/toolbar/insert-buttons.jsx
@@ -1,3 +1,4 @@
+import { TOOLBARS } from 'wix-rich-content-common';
 import { getDefault } from './../image-component';
 import { InsertPluginIcon } from '../icons';
 
@@ -7,6 +8,7 @@ export default ({ helpers, t }) => {
       type: 'file',
       name: 'Image',
       tooltipText: t('ImagePlugin_InsertButton_Tooltip'),
+      toolbars: [TOOLBARS.FOOTER, TOOLBARS.SIDE],
       Icon: InsertPluginIcon,
       componentData: getDefault(),
       helpers,