+++ packages/plugin-video/src/toolbar/insert-buttons.jsx
@@ -1,5 +1,5 @@
 import { DEFAULTS } from '../video-component';
-import { getModalStyles } from 'wix-rich-content-common';
+import { getModalStyles, TOOLBARS } from 'wix-rich-content-common';
 import VideoURLInputModal from './videoURLInputModal';
 import { InsertPluginIcon } from '../icons';
 
@@ -11,6 +11,7 @@ export default ({ helpers, t }) => {
       tooltipText: t('VideoPlugin_InsertButton_Tooltip'),
       Icon: InsertPluginIcon,
       componentData: DEFAULTS,
+      toolbars: [TOOLBARS.FOOTER, TOOLBARS.SIDE],
       modalElement: VideoURLInputModal,
       modalStyles: getModalStyles({ fullScreen: false }),
       helpers,