+++ CHANGELOG.md
@@ -5,8 +5,35 @@
 > - :rocket:     [New Feature]
 > - :bug:        [Bug Fix]
 > - :house:      [Internal]
+> - :book:       [Documentation]
 
 ## [Unreleased]
+
+## 1.3.0 (Aug 14, 2018)
+
+#### :boom: Breaking Change
+* `editor`
+  * `RichContentEditor`'s `alwaysShowSideToolbar`, `sideToolbarOffset`, `hideFooterToolbar` props removed
+* `general`
+  * `InsertButtons` API: `addToSideToolbar` property removed; `toolbars` property is required
+
+#### :rocket: New Feature
+* `editor`
+  * RCE `config.getToolbarSettings` API allows to customize toolbar instantiation, visibility, offset point, and buttons. Check [documentation](https://github.com/wix-incubator/rich-content/blob/develop/docs/ToolbarCustomization.md) for more details
+
+#### :bug: Bug Fix
+* `plugin-gallery`
+  * Add Media and Replace button UI and functionality
+  * Image Settings header is clickable
+  
+#### :house: Internal
+* `general`
+  * Published to public npm registry
+
+#### :book: Documentation
+* [Toolbar Customization](https://github.com/wix-incubator/rich-content/blob/develop/docs/ToolbarCustomization.md) added
+
+<br/>
 
 ## 1.2.14 (Aug 8, 2018)